Try to automatically do the "sentinel" optimization
This gets rid of the hand-crafted "textStyleSentinels" rule and instead automatically computes the set of characters that means "it's possible for this rule to match if you're looking at this character", and it uses that to optimize Choice. It adds a ton of complexity and was surprisingly hard to get right so need to evalute if it's worth keeping.